* common-src/util.c: Typo.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/branches/3_3@4924 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/util.c
@@ -1083,7 +1083,7 @@ int copy_file(
     }
 
     while((nb=read(infd, &buf, SIZEOF(buf))) > 0) {
-	if(full_write(outfd,&buf,nb) < nb) {
+	if(full_write(outfd,&buf,nb) < (size_t)nb) {
 	    save_errno = errno;
 	    quoted = quote_string(dst);
 	    *errmsg = vstrallocf(_("Error writing to '%s': %s"),
